refactor(desktop): split api.ts into types + runtime + mock halves (<500 each)
api.ts was 1,150 lines. Break it into focused files, no behavior change:

- apiTypes.ts (414): all DTOs + the Backend method contract
- apiEvents.ts (37): WailsRuntime types + browser-dev event-bus shim + mockEmit
- api.ts (114): realBackend detection, stream helpers, backend Proxy
- apiMockData.ts (67): mockMessages + a mutable md container (shared, so both
  mock halves can reassign state across module boundaries)
- apiMockA.ts (285) + apiMockB.ts (272): the mock methods, split in two
- apiMock.ts (7): merges the halves into mockBackend

Mock is exercised end-to-end by the Playwright suite: 34 e2e green, plus 61
unit, tsc + build clean.
