+++ desktop/frontend/src/apiMockA.ts
@@ -1,0 +1,285 @@
+// Mock backend — part A (Ready/init, mailbox, labels, attachments, AI, prompts,
+// bulk). Browser-dev only; merged with part B in apiMock.ts.
+import type { MessageDetail, Backend } from "./apiTypes";
+import { DEFAULT_KEYMAP } from "./apiTypes";
+import { md, mockMessages } from "./apiMockData";
+
+export const mockA: Partial<Backend> = {
+  async Ready() {
+    return true;
+  },
+  async InitError() {
+    return "";
+  },
+  async NeedsCredentials() {
+    return false;
+  },
+  async CredentialsPath() {
+    return "~/.config/giztui/credentials.json";
+  },
+  async ImportCredentials() {
+    return ""; // mock: no native file dialog in the browser
+  },
+  async RetryInit() {
+    /* mock: nothing to retry */
+  },
+  async Quit() {
+    /* mock: no-op in the browser */
+  },
+  async AccountEmail() {
+    return md.activeAccount === "work"
+      ? "you@company.com (mock)"
+      : "you@example.com (mock)";
+  },
+  async ListInbox(_pageToken: string, pageSize: number) {
+    return { messages: mockMessages.slice(0, pageSize || 50), nextPageToken: "" };
+  },
+  async Search(query: string, _pageToken: string, pageSize: number) {
+    const q = query.toLowerCase();
+    const filtered = mockMessages.filter(
+      (m) => m.subject.toLowerCase().includes(q) || m.from.toLowerCase().includes(q),
+    );
+    return { messages: filtered.slice(0, pageSize || 50), nextPageToken: "" };
+  },
+  async GetMessage(id: string) {
+    const m = mockMessages.find((x) => x.id === id) ?? mockMessages[0];
+    const plain = `${m.snippet}\n\nHi there,\n\nThis is the full body of "${m.subject}" rendered in the reading pane. In the packaged Wails app this content comes straight from Gmail via the GizTUI service layer.\n\nSee https://example.com/expenses for details (or www.example.org).\n\nBest,\n${m.from}`;
+    // Give some messages an HTML body so the HTML renderer is demonstrable.
+    const html =
+      Number(id.replace(/\D/g, "") || "0") % 2 === 0
+        ? `<div style="font-family:Arial,sans-serif">
+             <img src="https://example.com/logo.png" alt="logo" width="120">
+             <img src="cid:inlineimg1" alt="inline attachment">
+             <h2 style="color:#1a56db">${m.subject}</h2>
+             <p>Hi there,</p>
+             <p>This is a <strong>rich HTML</strong> version of the email, with
+             <a href="https://example.com">a link</a>, a list:</p>
+             <ul><li>First point</li><li>Second point</li><li>Third point</li></ul>
+             <p style="background:#f0f4ff;padding:12px;border-radius:8px">
+               A highlighted callout box rendered from the email's own inline styles.</p>
+             <p>Best,<br>${m.from}</p>
+           </div>`
+        : "";
+    return { ...m, to: "you@example.com", cc: "", plainText: plain, html };
+  },
+  async Archive() {},
+  async Trash() {},
+  async MarkRead() {},
+  async MarkUnread() {},
+  async Unarchive() {},
+  async Untrash() {},
+  async BulkUnarchive() {},
+  async BulkUntrash() {},
+  async ListLabels() {
+    return [
+      { id: "1", name: "Work" },
+      { id: "2", name: "Personal" },
+      { id: "3", name: "Travel" },
+    ];
+  },
+  async AIEnabled() {
+    return true;
+  },
+  async Summarize(id: string) {
+    const m = mockMessages.find((x) => x.id === id) ?? mockMessages[0];
+    await new Promise((r) => setTimeout(r, 500));
+    return `• ${m.subject}\n• Sent by ${m.from}\n• (mock summary) In the packaged app this is generated by your configured LLM via GizTUI's AIService.`;
+  },
+  async SendMail() {
+    await new Promise((r) => setTimeout(r, 300));
+  },
+  async Reply() {
+    await new Promise((r) => setTimeout(r, 300));
+  },
+  async MessageLabelIDs(id: string) {
+    return md.appliedLabels[id] ?? ["1"];
+  },
+  async ApplyLabel(messageID: string, labelID: string) {
+    const cur = new Set(md.appliedLabels[messageID] ?? ["1"]);
+    cur.add(labelID);
+    md.appliedLabels[messageID] = [...cur];
+  },
+  async RemoveLabel(messageID: string, labelID: string) {
+    const cur = new Set(md.appliedLabels[messageID] ?? ["1"]);
+    cur.delete(labelID);
+    md.appliedLabels[messageID] = [...cur];
+  },
+  async ListAttachments(id: string) {
+    if (id.endsWith("3") || id.endsWith("0")) {
+      return [
+        {
+          attachmentId: "att1",
+          filename: "invoice-2043.pdf",
+          mimeType: "application/pdf",
+          size: 84213,
+          type: "document",
+          inline: false,
+          contentId: "",
+        },
+        {
+          attachmentId: "att-inline",
+          filename: "image.png",
+          mimeType: "image/png",
+          size: 15210,
+          type: "image",
+          inline: true,
+          contentId: "inlineimg1",
+        },
+      ];
+    }
+    return [];
+  },
+  async DownloadAttachment(_m: string, _a: string, filename: string) {
+    await new Promise((r) => setTimeout(r, 300));
+    return `~/Downloads/gmail-attachments/${filename}`;
+  },
+  async SummarizeStream(id: string, force: boolean) {
+    // In the browser mock the streaming helper drives token delivery; this is
+    // only the fallback that returns the full text.
+    const s = await this.Summarize!(id);
+    return force ? s + "\n\n(regenerated)" : s;
+  },
+  async GenerateReply(id: string) {
+    const m = mockMessages.find((x) => x.id === id) ?? mockMessages[0];
+    await new Promise((r) => setTimeout(r, 400));
+    return `Hi,\n\nThanks for your message about "${m.subject}". (mock AI draft) In the packaged app this reply is drafted by your configured LLM — edit it before sending.\n\nBest,`;
+  },
+  async TouchUp(id: string) {
+    const m = mockMessages.find((x) => x.id === id) ?? mockMessages[0];
+    await new Promise((r) => setTimeout(r, 400));
+    return `## ${m.subject}\n\nHi there,\n\nThis is the message body, **reformatted** by the AI for readability — tidy paragraphs, fixed wrapping and clean markdown.\n\n- Point one\n- Point two\n\nBest,\n${m.from}`;
+  },
+  async MoveToLabel() {
+    await new Promise((r) => setTimeout(r, 200));
+  },
+  async BulkMoveToLabel() {
+    await new Promise((r) => setTimeout(r, 250));
+  },
+  async BulkArchive() {
+    await new Promise((r) => setTimeout(r, 250));
+  },
+  async BulkTrash() {
+    await new Promise((r) => setTimeout(r, 250));
+  },
+  async BulkMarkRead() {
+    await new Promise((r) => setTimeout(r, 200));
+  },
+  async BulkMarkUnread() {
+    await new Promise((r) => setTimeout(r, 200));
+  },
+  async BulkApplyLabel() {
+    await new Promise((r) => setTimeout(r, 200));
+  },
+  async BulkRemoveLabel() {
+    await new Promise((r) => setTimeout(r, 200));
+  },
+  async PromptsEnabled() {
+    return true;
+  },
+  async ListPrompts() {
+    return md.prompts.map((p) => ({
+      id: p.id,
+      name: p.name,
+      description: p.description,
+      category: p.category,
+    }));
+  },
+  async GetPrompt(id: number) {
+    return (
+      md.prompts.find((p) => p.id === id) ?? md.prompts[0]
+    );
+  },
+  async CreatePrompt(name: string, description: string, text: string, category: string) {
+    const id = Math.max(0, ...md.prompts.map((p) => p.id)) + 1;
+    md.prompts = [...md.prompts, { id, name, description, text, category }];
+    return id;
+  },
+  async UpdatePrompt(id: number, name: string, description: string, text: string, category: string) {
+    md.prompts = md.prompts.map((p) =>
+      p.id === id ? { id, name, description, text, category } : p,
+    );
+  },
+  async DeletePrompt(id: number) {
+    md.prompts = md.prompts.filter((p) => p.id !== id);
+  },
+  async RefinePromptText(text: string) {
+    await new Promise((r) => setTimeout(r, 400));
+    return `${text.trim()}\n\nBe concise and use {{body}} for the email content. (refined by mock AI)`;
+  },
+  async ApplyBulkPromptStream(ids: string[], _promptID: number) {
+    await new Promise((r) => setTimeout(r, 400));
+    return `Applied to ${ids.length} messages:\n\n• Combined key points across the selection\n• (mock bulk prompt result)`;
+  },
+  async ApplyPromptStream(_id: string, promptID: number, _force?: boolean) {
+    const names: Record<number, string> = {
+      1: "• Key point one\n• Key point two\n• Key point three",
+      2: "1. Reply to Ada by Friday (owner: you)\n2. Review the Q3 roadmap draft",
+      3: "Hi,\n\nThanks for the update — Friday works for me. See you then!\n\nBest,",
+      4: "Hola,\n\nEste es el cuerpo del correo traducido al español por tu LLM.",
+    };
+    return names[promptID] ?? "(mock prompt result)";
+  },
+  async CachedPrompts(_id: string) {
+    return []; // mock: no persisted results in the browser
+  },
+  async ListAccounts() {
+    return [
+      { id: "personal", email: "you@example.com", displayName: "Personal", active: md.activeAccount === "personal" },
+      { id: "work", email: "you@company.com", displayName: "Work", active: md.activeAccount === "work" },
+    ];
+  },
+  async SwitchAccount(id: string) {
+    await new Promise((r) => setTimeout(r, 250));
+    md.activeAccount = id;
+  },
+  async KeyMap() {
+    return DEFAULT_KEYMAP;
+  },
+  async ThreadingEnabled() {
+    return true;
+  },
+  async GetThread(_id: string) {
+    const mk = (i: number, from: string, unread: boolean): MessageDetail => ({
+      id: `t${i}`,
+      threadId: "thread-1",
+      subject: "Re: Project roadmap Q3",
+      from,
+      to: "you@example.com",
+      cc: "",
+      date: new Date(Date.now() - (3 - i) * 3600_000).toISOString(),
+      unread,
+      labels: ["Work"],
+      plainText:
+        i === 0
+          ? "Hi team,\n\nHere is the first pass at the Q3 roadmap. Thoughts?"
+          : i === 1
+            ? "Looks great — I'd move the analytics milestone earlier though."
+            : "Agreed. Let's lock it in for the Friday review.",
+      html: "",
+    });
+    return [
+      mk(0, "Ada Lovelace <ada@compute.org>", false),
+      mk(1, "you <you@example.com>", false),
+      mk(2, "Grace Hopper <grace@navy.mil>", true),
+    ];
+  },
+  async ThreadSummaryStream() {
+    return "• Ada shared the Q3 roadmap draft\n• Team agreed to move the analytics milestone earlier\n• Locked in for the Friday review";
+  },
+  async SavedQueriesEnabled() {
+    return true;
+  },
+  async ListSavedQueries() {
+    return md.queries;
+  },
+  async SaveQuery(name: string, query: string) {
+    md.queries = [
+      { id: md.queries.length + 1, name, query, description: "", category: "" },
+      ...md.queries,
+    ];
+  },
+  async DeleteSavedQuery(id: number) {
+    md.queries = md.queries.filter((q) => q.id !== id);
+  },
+  async RecordQueryUse() {},
+};
